pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
    "pytesseract>=0.3.13",
]
